refactor: split Contact dialog
