+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Contact/index.tsx
@@ -1,0 +1,3 @@
+export { DashboardContactDialog } from './Contact'
+export { DashboardContactDeleteConfirmDialog } from './DeleteConfirm'
+export { DashboardContactSearchDialog } from './Search'
