+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Contact/types.tsx
@@ -1,0 +1,5 @@
+import type { Profile } from '../../../../database'
+
+export interface ContactProps {
+    contact: Profile
+}
